requirements.txt
cdflib==1.3.3
contourpy==1.3.1
cycler==0.12.1
DateTime==5.5
fonttools==4.56.0
kiwisolver==1.4.8
matplotlib==3.10.1
numexpr==2.10.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
pyparsing==3.2.2
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
scipy==1.15.2
six==1.17.0
statsmodels==0.14.4
tzdata==2025.2
zope.interface==7.2
